pubmed-literature-mining/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
xata>=1.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dateutil>=2.8.2
numpy>=1.24.0
pandas>=2.0.0
regex>=2023.6.3

